chapters/ch3-opentelemetry/pyproject.toml
[project]
name = "ch3-opentelemetry"
version = "0.1.0"
description = "Automatic Tracing with OpenTelemetry in Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-distro>=0.46b0",
    "opentelemetry-exporter-otlp>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
]
